pyproject.toml
[tool.poetry]
name = "VocabVoyage"
version = "0.1.0"
description = ""
authors = ["ianycchen <ian@cloud-latitude.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.40.0"
langchain = "^0.3.7"
langgraph = "^0.2.45"
openai = "^1.54.3"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.5"
langchain-openai = "^0.2.6"
plotly = "^5.24.1"
graphviz = "^0.20.3"
ipython = "^8.29.0"
beautifulsoup4 = "^4.12.3"
chromadb = "^0.5.18"
langchain-chroma = "^0.1.4"
pypdf = "^5.1.0"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pillow = "^11.0.0"
google-cloud-aiplatform = "^1.72.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
